AgentAI now requires agent type registartion
Modified AgentAI so that it gets the AIChallenges list from the registered Ai, rather than receiving it as a parameter.
This forces all AI that use this system to be registered, thus ensuring that all mods have access to all registered AI and their AIChallenges.
AgentAI will now operate even with empty AIChallenge lists.