Wonder Spawning Restructuring
I've created a new WonderData class, which takes the wonder's `Type`, `int Priority`, and `bool Unique`.
The hook for adding a wonder to the spawn list now returns a list of WonderData, rather than a raw Type list.
If `Unique` is true, only one instance of it will ever be spawned.
The `Priority` can have four values:
- `0`: Will not be spawned. The same as not adding it to the list at all.
- `1`: Spawns as normal, if selected.
- `2`: These are  prioritised over the `Priority 1` wonders.
- `3` (Forced): These will always spawn first, even if this results in more wonders than the target spawn count. In such a case, no other wonders will ever spawn.

And I've added priority options for all three base game wonders  that ranges from 0 to 2.